backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
langchain==0.1.0
langgraph==0.0.18
langfuse==2.10.0
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.2
numpy==1.26.1
python-dotenv==1.0.0
twilio==8.9.1
httpx==0.25.1
anyio==3.7.1
loguru==0.7.2
jinja2==3.1.2
motor==3.3.2
pymongo==4.6.3
python-socketio[asyncio_server]==5.10.0
email-validator==2.2.0
python-multipart==0.0.12
